fix(backend): ProductOut.cost 改为float避免SQLite Numeric尾零
SQLite Numeric字段读回时Decimal带固定小数位（如7.6230000000），前端显示为长串零。
改为float后JSON序列化自然去尾零（7.623），前端无需处理。

+++ backend/app/schemas.py
@@ -8,7 +8,7 @@ class ProductOut(BaseModel):
     name: str | None = None
     spec: str | None = None
     category: str | None = None
-    cost: Decimal | None = None
+    cost: float | None = None  # float → JSON 不带尾零（如7.623而非7.6230000000）
 
     class Config:
         from_attributes = True
